New blog post about a trick I learned in Rust.
Across mutliple modules I needed to define functions
that return all, some, or derivative fields of a struct.
As these functions are different for each module, but need
to adhere to a common trait I needed a trick to make them
usable in same way somewhere else. The post describes the
solution.